fix: cast dummy covariates to float

+++ src/agricola/_internal/cli.py
@@ -189,7 +189,9 @@ def load_pheno_and_covars(
         df_covar_noid = df_covar.drop("IID", axis=1).drop(
             "FID", axis=1, errors="ignore"
         )
-        X = jnp.asarray(pd.get_dummies(df_covar_noid, columns=catcovariates).to_numpy())
+        X = jnp.asarray(
+            pd.get_dummies(df_covar_noid, columns=catcovariates, dtype=float).to_numpy()
+        )
     else:
         X = None
